refactor(generator): extract SymbolHelpers from SymbolDiscovery
